Moved dso_api..PatchedBrowsableAPIRenderer => rest_framework_dso.renderers
This functionality is generic, and easier to follow if there is just one
file with all renderers. This isn't really project-specific.
