+++ src/rest_framework_dso/renderers.py
@@ -3,8 +3,8 @@ from typing import Optional
 
 import orjson
 from gisserver.geometries import WGS84
+from rest_framework import renderers
 from rest_framework.relations import HyperlinkedRelatedField
-from rest_framework.renderers import JSONRenderer
 from rest_framework.serializers import ListSerializer, Serializer, SerializerMethodField
 from rest_framework.utils.serializer_helpers import ReturnDict, ReturnList
 from rest_framework_csv.renderers import CSVStreamingRenderer
@@ -37,7 +37,33 @@ class RendererMixin:
         self.paginator = paginator
 
 
-class HALJSONRenderer(RendererMixin, JSONRenderer):
+class BrowsableAPIRenderer(RendererMixin, renderers.BrowsableAPIRenderer):
+    def get_context(self, data, accepted_media_type, renderer_context):
+        context = super().get_context(data, accepted_media_type, renderer_context)
+
+        # Fix response content-type when it's filled in by the exception_handler
+        response = renderer_context["response"]
+        if response.content_type:
+            context["response_headers"]["Content-Type"] = response.content_type
+
+        return context
+
+    def render(self, data, accepted_media_type=None, renderer_context=None):
+        ret = super().render(
+            data,
+            accepted_media_type=accepted_media_type,
+            renderer_context=renderer_context,
+        )
+
+        # Make sure the browsable API always returns text/html
+        # by default it falls back to the current media,
+        # unless the response (e.g. exception_handler) has overwritten the type.
+        response = renderer_context["response"]
+        response["content-type"] = "text/html; charset=utf-8"
+        return ret
+
+
+class HALJSONRenderer(RendererMixin, renderers.JSONRenderer):
     media_type = "application/hal+json"
 
     # Define the paginator per media type.
@@ -85,7 +111,7 @@ class CSVRenderer(RendererMixin, CSVStreamingRenderer):
         yield from _chunked_output(output)
 
 
-class GeoJSONRenderer(RendererMixin, JSONRenderer):
+class GeoJSONRenderer(RendererMixin, renderers.JSONRenderer):
     """Convert the output into GeoJSON notation."""
 
     unlimited_page_size = True
